Add template for creating How-To task articles
To help users who want to contribute articles to the How-To, I created a template for tasks. I based it on the template created for Carina, but added some extra guidelines and How-To specific examples and links. I'll be asking everyone to review and provide feedback. When it is done, we can move it to the how-to repo and link to it from the contributing file.