Suggested edits to FAQ
Comments:

Suggest taking out "this is your role"  and replacing it with something like the text that I provided. The idea being is that the dev and support teams are uniquely qualified to provide this information, and if they write it down, we'll help make it great.  

Another thing that information developers provide is the "first user" perspective -- if it doesn't make sense, we ask questions and go back and forth on reviews until the content is clear.  

We haven't refined the workflow for the content lifecycle in a holistic way. When we do that, might be helpful to include link to that information, so that the dev and support people understand how their content is served into the documentation product.  Understanding the time required to move content from development into production is key for them to incorporate the time and milestones into their product or service development schedules.